Implemented UserProfilePage in mobile app

+++ XamApp/XamApp/AppShell.xaml.cs
@@ -14,6 +14,7 @@ namespace XamApp
             InitializeComponent();
             Routing.RegisterRoute(nameof(AddRoomPage), typeof(AddRoomPage));
             Routing.RegisterRoute(nameof(ChatRoomPage), typeof(ChatRoomPage));
+            Routing.RegisterRoute(nameof(UserProfilePage), typeof(UserProfilePage));
         }
 
         private async void OnSignOutClicked(object sender, EventArgs e)
@@ -22,22 +23,28 @@ namespace XamApp
             if (user != null)
             {
                 IsBusy = true;
-                var url = "/api/Account/SignOut?deviceUsedId=" + user.DeviceUsedId;
-
-                var response = await HTTPClient.PostAsync<string>(null, url, null);
-                if (!response.IsSuccessStatusCode)
-                {
-                    string message = await HTTPClient.GetResponseError(response);
-                    message += "\n\nFor your security, you must please clear all app data if this persists.";
-                    await DisplayAlert("Sign Out Failed", message, "Ok");
-                }
-                else
-                {
-                    HTTPClient.Clear();
-                    await DataStore.Instance.DeleteUserAsync();
-                    await Shell.Current.GoToAsync("//" + nameof(SignInPage));
-                }
+                await DoSignOut(this, user);
                 IsBusy = false;
+            }
+        }
+
+        public static async Task DoSignOut(Page page, User user)
+        {
+            long deviceUsedId = user == null ? 0 : user.DeviceUsedId;
+            string url = "/api/Account/SignOut?deviceUsedId=" + deviceUsedId;
+
+            var response = await HTTPClient.PostAsync<string>(null, url, null);
+            if (!response.IsSuccessStatusCode && response.StatusCode != System.Net.HttpStatusCode.Unauthorized)
+            {
+                string message = await HTTPClient.GetResponseError(response);
+                message += "\n\nFor your security, you must please clear all app data if this persists.";
+                await page.DisplayAlert("Sign Out Failed", message, "Ok");
+            }
+            else
+            {
+                HTTPClient.Clear();
+                await DataStore.Instance.DeleteUserAsync();
+                await Shell.Current.GoToAsync("//" + nameof(SignInPage));
             }
         }
     }
